Allow bounds on type parameters in `sql_function!`
This was... harder than it should be. What I really want to do is this:
`<$($type_arg $(: $bound:ty)*),* $(,)*>`. However, this will fail, since
doing: `$type_arg: $bound` complains that we can't use a `ty` token
there.

We can use `path`, but we can't capture the separater. We can't do `$(:
$($bound:path)+*)*`, since `+` means something special and can't be used
as a separator. We also can't do anything like `$(: $($bound:path)
$(+)*)` either, since Rust doesn't allow `path` fragments to be followed
by `+`.

Hell, we can't even do a token munching macro where we match something
like `$type_arg:ident : $bound:path $(rest:tt)*`, since Rust won't let
us have a `path` fragment followed by `tt` either! We can't even just
capture the tokens we care about since `<>` is not an individual token
tree, so `<$($tokens:tt)*>` is ambiguous.

So we have to implement a manual parser for this... We set up a state
machine that goes through and parses the whole thing one token a time.
I've tried to at least make it somewhat easier to follow than this has
been in the past by labeling each state, and the expected flow between
them.

It's way more code than it should be, but at least it's easy to test in
isolation.