Refactor our `table!` parsing code
The old code was a bit of a mess, and was suffering from some serious
bitrot. This is a rewrite that simplifies the flow, and structures it
more in line with how we would write this today. As a side effect, this
will allow attributes other than `#[doc]` to be applied to both tables
and columns. Importantly, this allows `#[allow]` to be applied, so
warnings we haven't considered in the past can be handled by our users.

The new code allows certain "invalid" things to get through (Honestly
the old code was so hard to follow, I'm not sure if this was true before
or not). For example, we permit `use` to come after meta items, and
`foo.bar.baz` is interpreted as `bar.baz`. These seme relatively minor
in the long run though.

I would have loved to have simplified this further by pulling out the
table name super early, but we still can't do that since Rust still
considers `use` ambiguous with `ident` matchers.