Fix our handling of unsigned types on MySQL
We were sending the right bytes, but we weren't properly setting the
`is_unsigned` flag on the bind parameters, resulting in it always being
sent as signed. This results in a client error about overflow when
deserializing large values, and either a server error or incorrect
results when serializing them.

The fact that the tests for this all used `-1` should have been a huge
red flag, but it slipped through.

While changing the type of `Mysql::BindCollector` could technically be
considered a breaking change under the strictest definition, I do not
consider this to be one. We have stated in the past that folks should
not rely on the exact types used in `impl Backend`. And unlike changing
`RawValue`, there's literally no reason for anybody to have depended on
the exact value of this type, unless they're implementing their own
Connection.

If we really want to avoid any potential source incompatibility, we can
actually do this without the custom bind collector, if we have the
`ToSql` implementations for `Unsigned` add an extra byte at the end, and
check to see if the length of the buffer is 1 byte larger than the
expected length. However, this is super hacky and much more likely to
break applications than this approach.

Fixes #1745.
Close #1749.