Add an explicit error kind for failed `SERIALIZABLE` transactions
"serializable" is the strictest isolation level defined in the SQL
standard. Any two transactions at this isolation level executed
concurrently must have the same results, regardless of which order they
are actually executed in. The standard also defines this as the default
isolation level, but there is no backend I'm aware of that both properly
implements the serializable isolation level, and uses is it as the
default.

We're generally very conservative about adding error kinds to this enum,
and only do so when there's a demonstrable use case for handling it
programatically. I believe this is clearly the case for this error code,
since the correct response to receiving it is almost always to retry the
transaction.

The test for this is pretty straight forward. We create two
transactions. Each one counts half the table, and then inserts a row in
the other half. When run serially, one of the transactions will get 1
for its count, and the other will get 2. Which transaction gets which
result depends on the order they are run, so one of them will fail to
commit. Interestingly, PG is too conservative in determining whether the
dependency exists. If I replace `other_i` with `i`, the transactions are
no longer dependent, but PG still thinks they are.

Currently this error is only detected on PostgreSQL, as that is the only
backend that we support isolation levels in our query builder, and the
semantics of this isolation level are unclear on other backends (SQLite
claims its the default, but the behavior it describes is `REPEATABLE
READ`. MySQL appears to just make your transactions deadlock).

This is the first half of #1612