Allow associations to structs with lifetime parameters
This was a bit of a pain to implement... I had hoped to just have the
syntax be `#[belongs_to(User<'_>)]`, which works on nightly... but is
feature gated on stable. I think that might be going away in 1.30, but
I'm not 100% certain yet. Second attempt was `#[belongs_to("User<'_>")]`
which works on nightly, but is feature gated on stable.

So the API has to be `#[belongs_to(some_key = "User<'_>")]` for the time
being. I've gone with `parent` as the key. This also makes our parsing
logic much more complicated, since we're now handling multiple forms
potentially given in multiple places. I want to refactor meta handling
in general eventually, but for now a little bit of spaghetti is fine.

The bigger messiness comes from what happens after the parsing. Next we
need to go through the type, and replace all instances of `'_` with some
unique lifetime, and stick that on the generics that we use for the
impl. This would actually look a bit cleaner if we had the users provide
lifetime names, but I'd rather be consistent with Rust syntax, instead
of simplifying our impl.

Fixes #1827.