Refactor our handling of nested options
Today our code for nested options assumes that all nested options are
required, and that option not being there is an error. This is probably
because when I wrote that function, it was for
`#[changeset_options(treat_none_as_null = "true")]`, which is the only
case where a nested option is required...

All other uses of this function immediately called `.ok` on it, meaning
if the error came from the meta item being malformed rather than the
nested item not being there, we either gave a suboptimal error message,
or ignored the error entirely.

For example, if you wrote `#[postgres = "Foo"]` when you meant
`#[postgres(sql_type = "Foo")]`, you'd get an error saying "missing
required options" rather than one telling you that `#[postgres]` was
malformed.

This forces callers to handle the error from malformed syntax, and
allows future features to distinguish from "the place we looked is
malformed" from "it's not there".