Fix use-after free in SQLite adapter
In `Statement::drop` we are attempting to get the raw connection pointer
from `inner_statement` after it has already been freed. We only use the
pointer if finalizing the statement fails, which *shouldn't* ever
happen, so this is unlikely to have caused a major issue in the wild
(e.g. we will not be issuing a security advisory for this). However,
getting that pointer in the first place is still UB, so it needs to be
fixed.

Fixes #1966