Use `diesel_infer_schema` 1.4.0
This crate was deprecated in 1.3, and is intended to no longer be
maintained. Because it uses internal APIs from Diesel (that are "public"
so they can be used by `infer_schema` and `cli`, but are not subject to
semver), we locked the version of Diesel to `~1.3.0`, expecting that
Cargo would allow multiple versions of Diesel to be used as it has in
the past.

What we've learned since then, is that Cargo will only resolve to
multiple versions if they are semver incompatible. So
`diesel_infer_schema` 1.3 cannot be used if anything requires Diesel
1.4.

Since our tests still use this deprecated crate, we are left with two
options. Update infer_schema to use Diesel 1.4, or stop using that crate
in our tests. We definitely need to stop using it at some point, but
since infer_schema appears to work with Diesel 1.4 without any code
changes, we can punt that for now.