Change `sql_function!` to be a proc macro
`sql_function!` basically has to include a full parser for arbitrary
Rust function declarations in `macro_rules!`. This is extremely verbose,
error prone, and has confusing to follow logic.

By changing `sql_function!` to be a proc macro instead, we can
dramatically simplify all of this logic, and also move towards unifying
`sql_function!` with `no_arg_sql_function!` (which requires some
rejiggering, as well as adding support for `where` clauses)

I've left a `macro_rules!` macro in the main crate to keep the docs from
moving (I'm not entirely clear on whether proc macros can even show up
in the documentation at all).

This change shouldn't have any user visible effects, but there were a
few unexpected cases the popped up. The most interesting/weird are that
it changes the resolution of `include!` in doc comments (before this
commit it was always resolving from the file that `sql_function` was
defined in which is wrong). It also started correctly identifying that
several imports were unused.

The most likely change to affect users is that `sql_function!` can on
longer be used inside of a function. This will cause the compiler to
complain that proc macros cannot expand to statements, even though we
are not expanding to a statement. This is a compiler limitation, not one
that is inherent to this change, and it will go away when the compiler
fixes this. Defining a `sql_function!` inside of a function also had
very limited usefulness since 5435fef30a05c72a37d01229d7743df0c8eb9768
anyway, so I'm not as concerned about breakage from this.